pyproject.toml
[project]
name = "worldclone"
version = "0.1.0"
description = "Multi-agent simulation vs. LLM forecaster bake-off on geopolitical prediction markets."
requires-python = ">=3.11"
dependencies = [
  "httpx>=0.27",
  "pydantic>=2.7",
  "litellm>=1.50",
  "exa-py>=1.0",
  "pyyaml>=6.0",
  "python-dotenv>=1.0",
  "matplotlib>=3.8",
  "numpy>=1.26",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-asyncio>=0.23",
  "ruff>=0.6",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["worldclone"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP"]
ignore = ["E501"]
